backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pyyaml>=6.0
httpx>=0.28.0
exa-py>=1.0.0
feedparser>=6.0.0
trafilatura>=2.0.0
openai>=1.30.0
python-dotenv>=1.0.0
numpy>=1.26.0
scikit-learn>=1.5.0
youtube-transcript-api>=1.0.0
dnspython>=2.0.0
copilotkit>=0.1.78
